Clean up the config GUI stuff
